requirements.txt
beautifulsoup4==4.14.3
pillow==12.1.1
pywebview==6.1
pywin32==311
pywin32-ctypes==0.2.3
requests==2.32.5
pyperclip==1.11.0